+++ LICENSE
@@ -1,6 +1,4 @@
-MIT License
-
-Copyright (c) 2021 MLO-lab
+Copyright (c) 2018 The Python Packaging Authority
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
