+++ MANIFEST.in
@@ -1,0 +1,1 @@
+graft joanapy/joana_app
