--- src/MANIFEST.in
@@ -1,1 +1,0 @@
-graft joanapy/joana_app
